fix(contracts): address PR review feedback (55-24913)

+++ contracts/src/main/java/com/team28/booking/contracts/events/BookingCompletedEvent.java
@@ -1,4 +1,6 @@
 package com.team28.booking.contracts.events;
 
-public record BookingCompletedEvent(Long bookingId, Long userId, Long providerId, Double totalPrice) {
+import java.math.BigDecimal;
+
+public record BookingCompletedEvent(Long bookingId, Long userId, Long providerId, BigDecimal totalPrice) {
 }
