+++ contracts/src/main/java/com/team28/booking/contracts/events/PaymentCompletedEvent.java
@@ -1,4 +1,6 @@
 package com.team28.booking.contracts.events;
 
-public record PaymentCompletedEvent(Long invoiceId, Long bookingId, Double amount) {
+import java.math.BigDecimal;
+
+public record PaymentCompletedEvent(Long invoiceId, Long bookingId, BigDecimal amount) {
 }
