+++ contracts/src/main/java/com/team28/booking/contracts/events/PaymentRefundedEvent.java
@@ -1,4 +1,6 @@
 package com.team28.booking.contracts.events;
 
-public record PaymentRefundedEvent(Long invoiceId, Long bookingId, Double refundAmount) {
+import java.math.BigDecimal;
+
+public record PaymentRefundedEvent(Long invoiceId, Long bookingId, BigDecimal refundAmount) {
 }
